# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API WebKasir",
        "description": "Dokumentasi API untuk aplikasi WebKasir",
        "contact": {
            "email": "support@webkasir.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/mobile/list_produk": {
            "get": {
                "tags": [
                    "Produk"
                ],
                "summary": "Ambil daftar produk yang aktif",
                "description": "Endpoint ini digunakan untuk mengambil semua produk yang statusnya aktif.",
                "operationId": "da4de773817440fef797769fb434ae62",
                "responses": {
                    "200": {
                        "description": "Data produk berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Barang Berhasil Diambiil"
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "kode_barang": {
                                                        "type": "string",
                                                        "example": "BRG001"
                                                    },
                                                    "nama_barang": {
                                                        "type": "string",
                                                        "example": "Kopi Hitam"
                                                    },
                                                    "harga_barang": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 10000
                                                    },
                                                    "gambar_produk": {
                                                        "type": "string",
                                                        "format": "url",
                                                        "example": "http://webkasir.milyascube.site/storage/gambar.jpg"
                                                    },
                                                    "kategori_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "stok": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "Aktif"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-06T12:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-07-06T12:30:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data barang tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data barang tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mobile/list_kategori": {
            "get": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Ambil daftar kategori produk",
                "description": "Endpoint ini digunakan untuk mengambil seluruh data kategori yang tersedia.",
                "operationId": "17a3ed76daf623f28c65084283ae3d35",
                "responses": {
                    "200": {
                        "description": "Data kategori berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Kategori berhasil diambil"
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "kode_kategori": {
                                                        "type": "string",
                                                        "example": "KTG001"
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "Minuman"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data kategori tidak ada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data kategori tidak ada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}